refactor: decompose large modules, improve naming, extract domains
- lark/serialize.py: Extract Serialize, SerializeMemoizer, Enumerator from utils.py
- lark/options.py: Extract LarkOptions from lark.py
- lark/persistence.py: Extract cache/pickle logic from lark.py
- lark/patterns.py: Extract PatternStr, PatternRE, TerminalDef from lexer.py
- lark/grammar_builder.py: Extract GrammarBuilder from load_grammar.py
- lark/grammar_transforms.py: Extract transformer classes from load_grammar.py
- lark/parsers/forest_visitors.py: Extract ForestVisitor, ForestTransformer from earley_forest.py
- Rename: Transformer_NonRecursive → IterativeTransformer (with alias)
- Rename: Transformer_InPlace → InPlaceTransformer (with alias)
- Rename: Interpreter → TreeInterpreter (with alias)
- Rename: merge_transformers → compose_transformers (with alias)
- All backward-compatible imports preserved
